Use nint to round number of azimuthal bins to avoid any floating point rounding problems.
git-svn-id: https://repository.astro.ex.ac.uk/torus/trunk@1996 25464f21-6099-402c-b30a-640f1d40d38b

+++ torus/amr_mod.f90
@@ -4814,7 +4814,7 @@ IF ( .NOT. gridConverged ) RETURN
       total_mass = cellVolume(thisOctal, subcell)  * 1.d30
 
       if ( thisOctal%cylindrical ) then
-         n_bin_az = twoPi / thisOctal%dPhi
+         n_bin_az = nint(twoPi / thisOctal%dPhi)
          massPerCell = ( (twoPi * thisOctal%r * 1.0e10) / n_bin_az ) * ( ave_density ** (1.0/3.0) ) * ( amrlimitscalar**(2.0/3.0) )
       else
          massPerCell = amrlimitscalar
